requirements.txt
bcrypt==5.0.0
blinker==1.9.0
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.3.1
gunicorn==25.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pillow==12.1.1
python-dotenv==1.2.2
SQLAlchemy==2.0.46
Werkzeug==3.1.5
WTForms==3.2.1
